♻️ refactor: extract sandbox contract to pkg
